1. CS50’s Introduction to Programming with Python - Harvard University/project/requirements.txt
googlemaps
simplejson
pandas
lxml
tabulate
fpdf2
#--upgrade google-auth-oauthlib